+++ google-cloud-spanner/src/main/java/com/google/cloud/spanner/AbstractResultSet.java
@@ -49,7 +49,8 @@ abstract class AbstractResultSet<R> extends AbstractStructReader implements Resu
   interface Listener {
     /**
      * Called when transaction metadata is seen. This method may be invoked at most once. If the
-     * method is invoked, it will precede {@link #onError(SpannerException)} or {@link #onDone()}.
+     * method is invoked, it will precede {@link #onError(SpannerException,boolean)} or {@link
+     * #onDone(boolean)}.
      */
     void onTransactionMetadata(Transaction transaction, boolean shouldIncludeId)
         throws SpannerException;
@@ -161,7 +162,7 @@ abstract class AbstractResultSet<R> extends AbstractStructReader implements Resu
     }
 
     /** it requests the initial prefetch chunks from gRPC stream */
-    default void requestPrefetchChunks() {};
+    default void requestPrefetchChunks() {}
   }
 
   static double valueProtoToFloat64(com.google.protobuf.Value proto) {