+++ proto-google-cloud-spanner-admin-database-v1/src/main/java/com/google/spanner/admin/database/v1/CrontabSpecOrBuilder.java
@@ -30,7 +30,7 @@ public interface CrontabSpecOrBuilder
    * <pre>
    * Required. Textual representation of the crontab. User can customize the
    * backup frequency and the backup version time using the cron
-   * expression. The version time must be in UTC timzeone.
+   * expression. The version time must be in UTC timezone.
    *
    * The backup will contain an externally consistent copy of the
    * database at the version time. Allowed frequencies are 12 hour, 1 day,
@@ -53,7 +53,7 @@ public interface CrontabSpecOrBuilder
    * <pre>
    * Required. Textual representation of the crontab. User can customize the
    * backup frequency and the backup version time using the cron
-   * expression. The version time must be in UTC timzeone.
+   * expression. The version time must be in UTC timezone.
    *
    * The backup will contain an externally consistent copy of the
    * database at the version time. Allowed frequencies are 12 hour, 1 day,